Don't persist account funded status in agent.toml
